% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hetFilterResults1}
\alias{hetFilterResults1}
\title{Individual1: 21B02337260 Filtered Heterozygous Mutations}
\format{
VCFv4.2
}
\usage{
hetFilterResults1
}
\description{
Individual1: 21B02337260 Filtered Heterozygous Mutations
}
\examples{
\dontrun{
 hetFilterResults1
}
}
\keyword{datasets}
