% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hetFilterResults2}
\alias{hetFilterResults2}
\title{Individual2: 21B02337261 Filtered Heterozygous Mutations}
\format{
VCFv4.2
}
\usage{
hetFilterResults2
}
\description{
Individual2: 21B02337261 Filtered Heterozygous Mutations
}
\examples{
\dontrun{
 hetFilterResults2
}
}
\keyword{datasets}
