% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hetFilterResults3}
\alias{hetFilterResults3}
\title{Individual3: 21B02337262 Filtered Heterozygous Mutations}
\format{
VCFv4.2
}
\usage{
hetFilterResults3
}
\description{
Individual3: 21B02337262 Filtered Heterozygous Mutations
}
\examples{
\dontrun{
 hetFilterResults3
}
}
\keyword{datasets}
