% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{homFilterResults1}
\alias{homFilterResults1}
\title{Individual1: 21B02337260 Filtered Homozygous Mutations}
\format{
VCFv4.2
}
\usage{
homFilterResults1
}
\description{
Individual1: 21B02337260 Filtered Homozygous Mutations
}
\examples{
\dontrun{
 homFilterResults1
}
}
\keyword{datasets}
