% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{homFilterResults2}
\alias{homFilterResults2}
\title{Individual2: 21B02337261 Filtered Homozygous Mutations}
\format{
VCFv4.2
}
\usage{
homFilterResults2
}
\description{
Individual2: 21B02337261 Filtered Homozygous Mutations
}
\examples{
\dontrun{
 homFilterResults2
}
}
\keyword{datasets}
