% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{homFilterResults3}
\alias{homFilterResults3}
\title{Individual3: 21B02337262 Filtered Homozygous Mutations}
\format{
VCFv4.2
}
\usage{
homFilterResults3
}
\description{
Individual3: 21B02337262 Filtered Homozygous Mutations
}
\examples{
\dontrun{
 homFilterResults3
}
}
\keyword{datasets}
