% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVCF.R
\name{readVCF}
\alias{readVCF}
\title{Import the VCF file and divides it into three sections for convenience in analysis.}
\usage{
readVCF(filename)
}
\arguments{
\item{filename}{The file name of the VCF file wish to be read.}
}
\value{
VCF file that is divided into three sections. "The meta region", "The fix region",
and "The gt region". Each section can be analyzed separately.
}
\description{
A function that reads in VCF file and divides it into three sections for convenience in analysis.
"Meta information" section, "Fixed information" section, and "Genotype information" section.
}
\examples{
# Example 1:
# Import data in VCF(variant call format) file
# Note: to read VCF files, it requires installation of vcfR package
library("vcfR")
individual1 <- system.file("extdata", "individual1.vcf", package = "CHDNFilter")

# Read using readVCF()

sample1 <- readVCF(individual1)

# Example 2:
# Import data in VCF(variant call format) file
# Note: to read VCF files, it requires installation of vcfR package
library("vcfR")
individual2 <- system.file("extdata", "individual2.vcf", package = "CHDNFilter")

# Read using readVCF()

sample2 <- readVCF(individual2)

# Example 3:
# Import data in VCF(variant call format) file
# Note: to read VCF files, it requires installation of vcfR package
library("vcfR")
individual3 <- system.file("extdata", "individual3.vcf", package = "CHDNFilter")

# Read using readVCF()

sample3 <- readVCF(individual3)


}
\references{
\href{https://cran.r-project.org/web/packages/vcfR/index.html}{Link}
}
