% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample1}
\alias{sample1}
\title{VCF files of different individuals}
\format{
VCFv4.2
}
\usage{
sample1
}
\description{
Individual1: 21B02337260
}
\examples{
\dontrun{
 sample1
}
}
\keyword{datasets}
