% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfPlotter.R
\name{vcfPlotter}
\alias{vcfPlotter}
\title{Plot the VCF file.}
\usage{
vcfPlotter(filename)
}
\arguments{
\item{filename}{The file name of the VCF file wish to be plotted (has to be a vcfR object,
can be obtained as result of readVCF).}
}
\value{
VCF file plot on 3 different aspects: read depth, mapping
quality, and Phred-Scaled quality
}
\description{
A function that plots VCF file on 3 different aspects: read depth, mapping
quality, and Phred-Scaled quality
}
\examples{
# Example 1:
# Import data in VCF(variant call format) file
# Note: to read VCF files, it requires installation of vcfR package
library("vcfR")
individual1 <- system.file("extdata", "individual1.vcf", package = "CHDNFilter")
vcffile1 <- vcfR::read.vcfR(file = individual1)

# plot using vcfPlotter()
my_plots <- vcfPlotter(vcffile1)

}
\references{
"Knaus BJ, Grünwald NJ (2016). “VcfR: an R package to manipulate and visualize VCF format data.”
\href{https://cran.r-project.org/web/packages/vcfR/index.html}{Link}

vicky, et al.
"Finding the class of a variable in R” stackoverflow, May 22, 2016
\href{https://stackoverflow.com/questions/37378978/finding-the-class-of-a-variable-in-r}{Link}

"Save Plot in Data Object in Base R (Example)” statisticsglobe
\href{https://statisticsglobe.com/save-plot-in-data-object-in-base-r}{Link}
}
